move nodemanager to a file
